2d_enhancement/dcp_dehaze/requirements.txt
numpy>=1.20
opencv-python-headless>=4.5
Pillow>=9.0
